ValidFen has enough info to ascertain the lib. Use that.
Otherwise it's not as usable in those other cases. Although, now that
I'm looking at it, maybe this was undoing what you figured you needed to
do.  However, if ValidFen is going to take a wrapped FEN, then it has
enough information. If it wasn't going to take a wrapped FEN, then we
need more than one of them (one for each type).

My personal viewpoint is that ValidFen should take a wrapped one and the
places that use it should appropriately do the wrapping.

Looks like the places in app which were previously not compiling are now
compiling, but are still hard coding the GameLib.Chess() values.